feat: Updating my_robot package with amcl.launch, which will has 3 nodes necessary to make localization happen.
The amcl.launch package will open the map server node, amcl node, and the move base node. The map server node provides map data as a ROS service to other nodes. The amcl node connects the world with the robot. It takes the odometry and laser scan data to perform amcl localization. Lastly, the move base node helps navigate to other positions. Also in this commit, the teleop package is included for localizing manually.