feat: adding updated my_robot package where files include changes to parameters.
Parameters were changed not only in the amcl.launch file, but also in the Trajectory Planner yaml files. These parameters were tuned based off of how it did during testing. Testing included placing the 2D navigation goal somewhere along the map and the trajectory planner gives the robot a trajetory to go to.